pyproject.toml
[tool.poetry]
name = "imcntr"
version = "0.2.0"
description = "A package providing an API for imaging controller based on an Arduino Nano Every to communicate via serial port."
authors = ["Clemens Trunner"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
pyserial = ">=3.5"

[tool.poetry.dev-dependencies]
python-semantic-release = {version = "^9.20.0", python = "^3.9"}
myst-nb = "^1.2.0"
sphinx-autoapi = "^3.6.0"
sphinx-rtd-theme = "^3.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_toml =  ["pyproject.toml:tool.poetry.version",]
